+++ src/views/setting/remind/index.vue
@@ -103,7 +103,14 @@ export default {
         })
     },
     deleteAlarm(id) {
-      this.$store.dispatch('DeleteAlarmData', id)
+      this.$confirm('確定要刪除嗎？', '', {
+        confirmButtonText: '確定',
+        cancelButtonText: '取消'
+      })
+        .then(() => {
+          this.$store.dispatch('DeleteAlarmData', id)
+        })
+        .catch(() => {})
     },
     openDialog(rawData) {
       this.showDialog = true
